add memory limit and balooning
new option, ex.: --mem_limit=80 to only start new workers if not more than 80%
memory are used on this machine.

new option --backgrounding-threshold=20 to put current job into background mode
after 20seconds and start working on the next job.
The idea is that long running plugins usually wait on some kind of network response
or timeout and don't do much. So it's safe to start a new check in parallel.